fastapi/requirements.txt
fastapi
pillow
torch
torchvision
python-multipart
aiofiles